requirements.txt
streamlit==1.30.0
openai==1.109.1
openai[async]==1.109.1
mcp==1.15.0
langchain-mcp-adapters==0.1.10
langchain-openai==0.3.33
requests==2.32.5
